B2B project creation
The B2B world operates differently from the B2C or C2C world. Businesses
work with other businesses on credit. When a buyer business orders goods
from the seller business, the seller business issues an invoice for the same.
This invoice for the goods contains various information like the details of the
goods purchased and when it should be paid. This is known in accounting
terminology as “Accounts Receivable”.
Seller business interacts with various businesses and sells goods to all of them
at various times. Hence, the seller business needs to keep track of the total
amount it owes from all the buyers. This involves keeping track of all invoices
from all the buyers. Each invoice will have various important fields like a
payment due date, invoice date, invoice amount, baseline date etc.The buyer business needs to clear its amount due before the due date. However,
in real-world scenarios, the invoices are not always cleared ie. paid in full
amount by the due date. The date on which a customer clears the payment for
an invoice is called the payment date.